Reworking locking and unlock messages
No a JS-Handler watches the user not to leave the edit page and asks for confirmation, if that happens. If the user confirms, he will navigation away but also the lock is removed (clear decision made). With that feature it wont happen again, that you press on back and lose all your content accidently :)